requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements/requirements.txt requirements/requirements.in
#
certifi==2022.12.7
    # via requests
charset-normalizer==3.1.0
    # via requests
colorama==0.4.6
    # via halo
halo==0.0.31
    # via -r requirements/requirements.in
idna==3.4
    # via requests
log-symbols==0.0.14
    # via halo
python-dotenv==1.0.0
    # via -r requirements/requirements.in
requests==2.28.2
    # via -r requirements/requirements.in
six==1.16.0
    # via halo
spinners==0.0.24
    # via halo
termcolor==2.2.0
    # via
    #   -r requirements/requirements.in
    #   halo
urllib3==1.26.15
    # via requests
